Move tests for extractGeocodeOptions
